Moving Batcher V2 to existing batching package
 - Renamed to `v2.RequestBuilder` to `BatchingRequestBuilder`.
 - Adapted to existing BatchingSettings.
 - In case ElementCountThreshold or RequestByteCountThreshold is null, then using 0 as threshold level.
 - Defaulting to 1ms in case delayThreshold value is unset.
