+++ gax/src/main/java/com/google/api/gax/batching/BatcherImpl.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.batching.v2;
+package com.google.api.gax.batching;
 
 import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 
@@ -100,7 +100,10 @@ public class BatcherImpl<ElementT, ElementResultT, RequestT, ResponseT>
     Preconditions.checkNotNull(executor, "executor cannot be null");
     currentOpenBatch = new Batch<>(prototype, batchingDescriptor, batchingSettings);
 
-    long delay = batchingSettings.getDelayThreshold().toMillis();
+    long delay =
+        batchingSettings.getDelayThreshold() == null
+            ? 1L
+            : batchingSettings.getDelayThreshold().toMillis();
     PushCurrentBatchRunnable<ElementT, ElementResultT, RequestT, ResponseT> runnable =
         new PushCurrentBatchRunnable<>(this);
     scheduledFuture =
@@ -208,7 +211,7 @@ public class BatcherImpl<ElementT, ElementResultT, RequestT, ResponseT>
    * future results for one batch.
    */
   private static class Batch<ElementT, ElementResultT, RequestT, ResponseT> {
-    private final RequestBuilder<ElementT, RequestT> builder;
+    private final BatchingRequestBuilder<ElementT, RequestT> builder;
     private final List<SettableApiFuture<ElementResultT>> results;
     private final BatchingDescriptor<ElementT, ElementResultT, RequestT, ResponseT> descriptor;
     private final long elementThreshold;
@@ -223,9 +226,11 @@ public class BatcherImpl<ElementT, ElementResultT, RequestT, ResponseT>
         BatchingSettings batchingSettings) {
       this.descriptor = descriptor;
       this.builder = descriptor.newRequestBuilder(prototype);
-      this.elementThreshold = batchingSettings.getElementCountThreshold();
-      this.bytesThreshold = batchingSettings.getRequestByteThreshold();
       this.results = new ArrayList<>();
+      Long elementCountThreshold = batchingSettings.getElementCountThreshold();
+      this.elementThreshold = elementCountThreshold == null ? 0 : elementCountThreshold;
+      Long requestByteThreshold = batchingSettings.getRequestByteThreshold();
+      this.bytesThreshold = requestByteThreshold == null ? 0 : requestByteThreshold;
     }
 
     void add(ElementT element, SettableApiFuture<ElementResultT> result) {
