+++ gax/src/main/java/com/google/api/gax/batching/BatchingRequestBuilder.java
@@ -27,45 +27,27 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.batching.v2;
+package com.google.api.gax.batching;
 
-import com.google.api.core.ApiFuture;
 import com.google.api.core.BetaApi;
+import com.google.api.core.InternalExtensionOnly;
 
 /**
- * Represents a batching context where individual elements will be accumulated and flushed in a
- * large batch request at some point in the future. The buffered elements can be flushed manually or
- * when triggered by an internal threshold. This is intended to be used for high throughput
- * scenarios at the cost of latency.
+ * Adapter to pack individual elements into a larger batch request.
+ *
+ * <p>The implementation for this interface will be implemented by service specific client or auto
+ * generated by the gapic-generator.
  *
  * @param <ElementT> The type of each individual element to be batched.
- * @param <ElementResultT> The type of the result for each individual element.
+ * @param <RequestT> The type of the request that will contain the accumulated elements.
  */
 @BetaApi("The surface for batching is not stable yet and may change in the future.")
-public interface Batcher<ElementT, ElementResultT> extends AutoCloseable {
+@InternalExtensionOnly("For google-cloud-java client use only.")
+public interface BatchingRequestBuilder<ElementT, RequestT> {
 
-  /**
-   * Queues the passed in element to be sent at some point in the future.
-   *
-   * <p>The element will be sent as part of a larger batch request at some point in the future. The
-   * returned {@link ApiFuture} will be resolved once the result for the element has been extracted
-   * from the batch response.
-   *
-   * <p>Note: Cancelling returned result simply marks the future cancelled, It would not stop the
-   * batch request.
-   */
-  ApiFuture<ElementResultT> add(ElementT entry);
+  /** Adds element object into client specific batch request. */
+  void add(ElementT element);
 
-  /**
-   * Synchronously sends any pending elements as a batch and waits for all outstanding batches to be
-   * complete.
-   */
-  void flush() throws InterruptedException;
-
-  /**
-   * Closes this Batcher by preventing new elements from being added and flushing the existing
-   * elements.
-   */
-  @Override
-  void close() throws InterruptedException;
+  /** Returns the collected elements as a single client specific batch request. */
+  RequestT build();
 }
